Rename `PayPalNativeCheckout` to `PayPalNativePayments` (#114)
* Update build.gradle, docs and changelog

* Rename module

* Rename package name

* Fix typo

* Name package to paypalnativepayments

* Fix docs
