+++ .github/workflows/release.yml
@@ -81,7 +81,7 @@ jobs:
         with:
           module: PayPalWebPayments
   unit_test_paypal_native:
-    name: PayPal Native Unit Tests
+    name: PayPal Native Payments Unit Tests
     runs-on: ubuntu-latest
     steps:
       - name: Checkout Repository
@@ -94,7 +94,7 @@ jobs:
       - name: Run Unit Tests
         uses: ./.github/actions/unit_test_module
         with:
-          module: PayPalNativeCheckout
+          module: PayPalNativePayments
 
   unit_test_finished:
     needs: [unit_test_card, unit_test_core, unit_test_paypal_web, unit_test_paypal_native]
