+++ .github/workflows/static_analysis.yml
@@ -74,7 +74,7 @@ jobs:
             CorePayments/build/outputs/aar/CorePayments-release.aar
             CardPayments/build/outputs/aar/CardPayments-release.aar
             PayPalWebPayments/build/outputs/aar/PayPalWebPayments-release.aar
-            PayPalNativeCheckout/build/outputs/aar/PayPalNativeCheckout-release.aar
+            PayPalNativePayments/build/outputs/aar/PayPalNativePayments-release.aar
             PaymentButtons/build/outputs/aar/PaymentButtons-release.aar
 
       # Set up Diffuse
@@ -95,7 +95,7 @@ jobs:
         run: diffuse diff --aar main-aar/Core/build/outputs/aar/Core-release.aar current-aar/CorePayments/build/outputs/aar/CorePayments-release.aar
       - name: Run Diffuse Analysis - PayPalWebPayments
         run: diffuse diff --aar main-aar/PayPalWebCheckout/build/outputs/aar/PayPalWebCheckout-release.aar current-aar/PayPalWebPayments/build/outputs/aar/PayPalWebPayments-release.aar
-      - name: Run Diffuse Analysis - PayPalNativeCheckout
-        run: diffuse diff --aar main-aar/PayPalNativeCheckout/build/outputs/aar/PayPalNativeCheckout-release.aar current-aar/PayPalNativeCheckout/build/outputs/aar/PayPalNativeCheckout-release.aar
+      - name: Run Diffuse Analysis - PayPalNativePayments
+        run: diffuse diff --aar main-aar/PayPalNativeCheckout/build/outputs/aar/PayPalNativeCheckout-release.aar current-aar/PayPalNativePayments/build/outputs/aar/PayPalNativePayments-release.aar
       - name: Run Diffuse Analysis - PaymentButtons
         run: diffuse diff --aar main-aar/PayPalUI/build/outputs/aar/PayPalUI-release.aar current-aar/PaymentButtons/build/outputs/aar/PaymentButtons-release.aar
