+++ Demo/build.gradle
@@ -62,7 +62,7 @@ android {
 dependencies {
     debugImplementation project(':CardPayments')
     debugImplementation project(':PayPalDataCollector')
-    debugImplementation project(':PayPalNativeCheckout')
+    debugImplementation project(':PayPalNativePayments')
     debugImplementation project(':PayPalWebPayments')
     debugImplementation project(':PaymentButtons')
     debugImplementation project(':Venmo')
