+++ PayPalNativePayments/proguard-rules.pro
