+++ PayPalNativePayments/src/main/java/com/paypal/android/paypalnativepayments/PayPalNativeCheckoutClient.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.checkout
+package com.paypal.android.paypalnativepayments
 
 import android.app.Application
 import com.paypal.android.corepayments.API
