+++ PayPalNativePayments/src/test/java/com/paypal/android/paypalnativepayments/RandomValueGenerators.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.checkout
+package com.paypal.android.paypalnativepayments
 
 import java.util.UUID
