+++ PaymentButtons/src/main/java/com/paypal/android/paymentbuttons/error/Exceptions.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.checkout.paymentbutton.error
+package com.paypal.android.paypalnativepayments.paymentbutton.error
 
 internal fun createFormattedIllegalArgumentException(enumName: String, enumValues: Int): IllegalArgumentException {
     val exceptionMessage = "Attempted to create a $enumName with an invalid index. Please use an" +
