+++ docs/README.md
@@ -3,5 +3,5 @@
 - [CardPayments](CardPayments)
 - [PaymentButtons](PaymentButtons)
 - [PayPal Web Payments](PayPalWebPayments)
-- [PayPal Native Checkout](PayPalNativeCheckout)
-- [PayPal Native Checkout with Billing Agreement](PayPalNativeCheckout/README.md#billing-agreement)
+- [PayPal Native Payments](PayPalNativePayments)
+- [PayPal Native Payments with Billing Agreement](PayPalNativePayments/README.md#billing-agreement)
